Refactor scrollable board into component
