+++ src/components/catalog/ClubModal.tsx
@@ -1,5 +1,6 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useRef } from "react";
 import { IClub } from "@/lib/models/Club";
+import Board from "./Board";
 
 type ClubModalProps = {
   club: IClub;
@@ -8,21 +9,6 @@ type ClubModalProps = {
 
 const ClubModal = ({ club, onClose }: ClubModalProps) => {
   const modalRef = useRef<HTMLDivElement>(null);
-  const [isScrolledToStart, setIsScrolledToStart] = useState(true);
-  const [isScrolledToEnd, setIsScrolledToEnd] = useState(false);
-  const scrollContainerRef = useRef<HTMLDivElement>(null);
-
-  const handleScroll = () => {
-    if (!scrollContainerRef.current) return;
-
-    const { scrollLeft, scrollWidth, clientWidth } = scrollContainerRef.current;
-    setIsScrolledToStart(scrollLeft === 0);
-    setIsScrolledToEnd(scrollLeft + clientWidth >= scrollWidth);
-  };
-
-  useEffect(() => {
-    handleScroll();
-  }, []);
 
   useEffect(() => {
     const handleKeyDown = (event: KeyboardEvent) => {
@@ -78,38 +64,7 @@ const ClubModal = ({ club, onClose }: ClubModalProps) => {
               <div className="text-gray-700 text-md mt-4">Unknown affiliation</div>
             )}
             <div className="text-gray-700 mt-4">{club.description || "No description"}</div>
-            {club.leaders.length > 0 && (
-              <div className="mt-4">
-                <div className="text-xl font-bold">Board</div>
-                <div className="relative">
-                  <div
-                    className="flex flex-row gap-4 mt-2 overflow-x-auto whitespace-nowrap"
-                    ref={scrollContainerRef}
-                    onScroll={handleScroll}
-                  >
-                    {!isScrolledToStart && (
-                      <div className="absolute left-0 top-0 bottom-0 w-8 bg-gradient-to-l from-transparent to-white pointer-events-none"></div>
-                    )}
-                    {!isScrolledToEnd && (
-                      <div className="absolute right-0 top-0 bottom-0 w-12 bg-gradient-to-r from-transparent to-white pointer-events-none"></div>
-                    )}
-
-                    {club.leaders.map((leader, index) => (
-                      <div key={index} className="flex items-center rounded-lg border p-2 text-sm">
-                        <div className="flex flex-col items-start">
-                          <div className="flex flex-row justify-between w-full">
-                            <div className="text-md font-semibold">{leader.name}</div>
-                            {leader.year && <div className="">{"'" + (leader.year % 100)} </div>}
-                          </div>
-                          <div className="">{leader.email}</div>
-                          <div className="">{leader.role}</div>
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                </div>
-              </div>
-            )}
+            <Board leaders={club.leaders} />
           </div>
           <div className="flex flex-col">
             <div className="text-gray-700 mb-4">something</div>
